requirements.txt
Django>=3.2,<4.0
python-decouple
python-dotenv==1.0.1